Se modifica el Nombre del archivi ServiceProvider por error de conflicto por ActiveCampaignServiceProvider
